requirements.txt
-f /usr/share/pip-wheels
asgiref==3.2.7
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
blessed==1.17.5
botocore==1.15.49
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
colorthief==0.2.1
cryptography==2.9.2
cycler==0.10.0
Django==3.0.6
django-background-tasks==1.2.5
django-common-helpers==0.9.2
django-compat==1.0.15
django-cron==0.5.1
django-crontab==0.7.1
django-npm==1.0.0
django-npm-mjs==2.0.1
django-taggit==1.3.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
importlib-metadata==1.6.0
jmespath==0.10.0
JSON-minify==0.3.0
jsonschema==3.2.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.2.0
pycparser==2.20
PyNaCl==1.3.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
scipy==1.5.0
semantic-version==2.5.0
six==1.11.0
spotipy==2.12.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0
